Fix click on POI
