+++ src/panel/category_panel.js
@@ -6,6 +6,7 @@ import debounce from '../libs/debounce';
 import poiSubClass from '../mapbox/poi_subclass';
 import nconf from '@qwant/nconf-getter';
 import Telemetry from '../libs/telemetry';
+import layouts from 'src/panel/layouts.js';
 import ReactCategoryPanel from './category/CategoryPanel';
 
 const categoryConfig = nconf.get().category;
@@ -112,6 +113,8 @@ export default class CategoryPanel {
       hasError={!this.pois || this.pois.length === 0}
       zoomIn={!this.pois}
       close={this.closeAction}
+      selectPoi={this.selectPoi}
+      highlightPoiMarker={this.highlightPoiMarker}
     />;
     ReactDOM.render(panel, document.querySelector('.react_panel__container'));
   }
@@ -136,5 +139,43 @@ export default class CategoryPanel {
 
   removeCategoryMarkers() {
     fire('remove_category_markers', this.pois);
+  }
+
+  selectPoi = poi => {
+    const previousMarker = document.querySelector('.mapboxgl-marker.active');
+    if (previousMarker) {
+      previousMarker.classList.remove('active');
+    }
+    if (poi.meta && poi.meta.source) {
+      Telemetry.add('open', 'poi', poi.meta.source,
+        Telemetry.buildInteractionData({
+          id: poi.id,
+          source: poi.meta.source,
+          template: 'multiple',
+          zone: 'list',
+          element: 'item',
+          category: this.categoryName,
+        })
+      );
+    }
+    window.app.navigateTo(`/place/${poi.toUrl()}`, {
+      poi: poi.serialize(),
+      isFromCategory: true,
+      sourceCategory: this.categoryName,
+      layout: layouts.LIST,
+      centerMap: true,
+    });
+    this.highlightPoiMarker(poi, true);
+  }
+
+  highlightPoiMarker = (poi, highlight) => {
+    const marker = document.getElementById(poi.marker_id);
+    if (marker) {
+      if (highlight) {
+        marker.classList.add('active');
+      } else {
+        marker.classList.remove('active');
+      }
+    }
   }
 }
